rag-chatbot/requirements.txt
PyMuPDF
sentence-transformers
faiss-cpu
google-generativeai
flask
flaREMOVEDcors
